refactor(interaction): extract clipboard-probe.ts shared primitives from paste-image
Pure move, no behavior change: the reader-result contract
(ClipboardImageResult/FailureResult/FailureKind), ClipboardBackend, the
never-rejecting runTool with its ToolRun/FailedRun shapes and 3s timeout,
failureDetail, and the bounded sniffed readCopiedImage leave paste-image.ts
for a sibling internal module (src/updater precedent: bundled through the
paste-image entry, no exports-map change). paste-image.ts re-exports
ClipboardBackend/ClipboardImageResult so its public type surface is
unchanged. Groundwork for the win32/darwin native probes (S39, D55).
